==> 2_Data cleaning round 2 using Lau Pa Sat from TripAdvisor (2).ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.feature_extraction.text import CountVectorizer

# Recall that df contains unlemmatized reviews and df2 contains lemmatized reviews
df = pd.read_pickle("df.pkl")
df2 = pd.read_pickle("df2.pkl")

df.iloc[1241,:]

# I will co
def clean_rd_2(sentence):
    sentence = sentence.replace('aaa','')
    sentence = sentence.replace('aboutthen','about then')
    sentence = sentence.replace('accessiblesome','accessible some')
    sentence = sentence.replace('actuallythere','actually there')
    sentence = sentence.replace('adddition', 'addition')
    sentence = sentence.replace('addednow','added now')
    sentence = sentence.replace('adverstisement','advertisement')
    sentence = sentence.replace('affordablecons','affordable cons')
    sentence = sentence.replace('affordablegreat','affordable great')
    sentence = sentence.replace('affordablei','affordable i')
    sentence = sentence.replace('aforadable','affordable')
    sentence = sentence.replace('againlanguage','again language')
    sentence = sentence.replace('againthoroghly','again thoroghly')
    sentence = sentence.replace('agobut','ago but')
    sentence = sentence.replace('agoit','ago it')
    sentence = sentence.replace('airportthe','airport the')
    sentence = sentence.replace('alikethe','alike the')
    sentence = sentence.replace('allthe','all the')
    sentence = sentence.replace('allthen','all then')
    sentence = sentence.replace('allthis','all the')
    sentence = sentence.replace('allways','all ways')
    sentence = sentence.replace('alots','a lot')
    sentence = sentence.replace('alternativesi','alternatives i')
    sentence = sentence.replace('amaizing','amazing')
    sentence = sentence.replace('amazingthe','amazing the')
    sentence = sentence.replace('amazingwe','amazing we')
    sentence = sentence.replace('ambiance','ambience')
    sentence = sentence.replace('ambiancea','ambience a')
    sentence = sentence.replace('ambiancemany','ambience many')
    sentence = sentence.replace('ambrosiamany','ambrosia many')
    sentence = sentence.replace('americanindian','american indian')
    sentence = sentence.replace('areacautionary','area cautionary')
    sentence = sentence.replace('areagreat','area great')
    sentence = sentence.replace('areamost','area most')
    sentence = sentence.replace('areamost','area most')
    sentence = sentence.replace('areait','area it')
    sentence = sentence.replace('areanatural','area natural')
    sentence = sentence.replace('areaplenty','area plenty')
    sentence = sentence.replace('areathe','area the')
    sentence = sentence.replace('areatry','area try')
    sentence = sentence.replace('areawaitingwaitingdidnt','area waiting waiting did not')
    sentence = sentence.replace('areawould','area would')
    sentence = sentence.replace('arrivedthe','arrived the')
    sentence = sentence.replace('asiaa','asia')
    sentence = sentence.replace('asiamalaysia','asia malaysia')
    sentence = sentence.replace('asiaticindian','asia indian')
    sentence = sentence.replace('aspectgreat','aspect great')
    sentence = sentence.replace('atleast','at least')
    sentence = sentence.replace('atmosphereapart','atmosphere apart')
    sentence = sentence.replace('atmosphereno','atmosphere no')
    sentence = sentence.replace('atmospherethis','atmosphere this')
    sentence = sentence.replace('availiable','available')
    sentence = sentence.replace('availabilityi','availability i')
    sentence = sentence.replace('availablelookout','available lookout')
    sentence = sentence.replace('avoidednonetheless','avoided nonetheless')
    sentence = sentence.replace('awayloved','away loved')
    sentence = sentence.replace('awesomethe','awesome the')
    sentence = sentence.replace('awesomethose','awesome those')
    sentence = sentence.replace('awfulshop','awful shop')
    sentence = sentence.replace('awsome','awesome')
    sentence = sentence.replace('badnot','bad not')
    sentence = sentence.replace('bangkokhaggling','bangkok haggling')
    sentence = sentence.replace('barsjust','bars just')
    sentence = sentence.replace('barsthis','bars this')
    sentence = sentence.replace('bc', 'becuase')
    sentence = sentence.replace('bcos','because')
    sentence = sentence.replace('bebring','be bring')
    sentence = sentence.replace('beeen','been')
    sentence = sentence.replace('beefchoice','beef choice')
    sentence = sentence.replace('beefballs','beef balls')
    sentence = sentence.replace('beefmy','beef my')
    sentence = sentence.replace('beerat','beer at')
    sentence = sentence.replace('beerbest','beer best')
    sentence = sentence.replace('beerdrinks','beer drinks')
    sentence = sentence.replace('beerfab','beer fab')
    sentence = sentence.replace('beeri','beer i')
    sentence = sentence.replace('beernice','beer nice')
    sentence = sentence.replace('beernot','beer not')
    sentence = sentence.replace('beeror','beer or')
    sentence = sentence.replace('beersby','beers by')
    sentence = sentence.replace('beersthe','beers the')
    sentence = sentence.replace('beerthe','beer the')
    sentence = sentence.replace('beervery','beer very')
    sentence = sentence.replace('beoriginally','be originally')
    sentence = sentence.replace('bestaround','best around')
    sentence = sentence.replace('besteating','best eating')
    sentence = sentence.replace('bestsatays','best satays')
    sentence = sentence.replace('betterdont','better do not')
    sentence = sentence.replace('bettertoo','better too')
    sentence = sentence.replace('betterusually','better usually')
    sentence = sentence.replace('bewarethe','beware the')
    sentence = sentence.replace('blendedgreat','blended great')
    sentence = sentence.replace('bonusand','bonus and')
    sentence = sentence.replace('boomingthe','booming the')
    sentence = sentence.replace('breakfastbakery','breakfast bakery')
    sentence = sentence.replace('breakfastmany','breakfast many')
    sentence = sentence.replace('breathtakingservice','breathtaking service')
    sentence = sentence.replace('budgethowever','budget however')
    sentence = sentence.replace('buildingsan','buildings an')
    sentence = sentence.replace('buildingsfood',' my')
    sentence = sentence.replace('buildingswhilst','buildings whilst')
    sentence = sentence.replace('busines','business')
    sentence = sentence.replace('busthe','bus the')
    sentence = sentence.replace('cakesrice','cakes rice')
    sentence = sentence.replace('cakesnoodles','cakes noodles')
    sentence = sentence.replace('cakesome','cake some')
    sentence = sentence.replace('cakeyou','cake you')
    sentence = sentence.replace('carefull','careful')
    sentence = sentence.replace('cbda','cbd a')
    sentence = sentence.replace('cbdevening','cbd evening')
    sentence = sentence.replace('cbdthere','cbd there')
    sentence = sentence.replace('centrefood','centre food')
    sentence = sentence.replace('centerprices','center prices')
    sentence = sentence.replace('centersi','centers i')
    sentence = sentence.replace('changethe','change the')
    sentence = sentence.replace('cheapa','cheap a')
    sentence = sentence.replace('cheapdont','cheap do not')
    sentence = sentence.replace('cheapeateries','cheap eateries')
    sentence = sentence.replace('cheapenjoy','cheap enjoy')
    sentence = sentence.replace('cheaperand','cheaper and')
    sentence = sentence.replace('cheaperthe','cheaper the')
    sentence = sentence.replace('cheapif','cheap if')
    sentence = sentence.replace('cheaplong','cheap long')
    sentence = sentence.replace('chickedn','chicken')
    sentence = sentence.replace('chickenbeef','chicken beef')
    sentence = sentence.replace('chickenmuttonbeef','chicken mutton beef')
    sentence = sentence.replace('chickennot','chicken not')
    sentence = sentence.replace('chickenporkprawns','chicken pork prawns')
    sentence = sentence.replace('chickenprawn','chicken prawn')
    sentence = sentence.replace('chiken','chicken')
    sentence = sentence.replace('chillie','chili')
    sentence = sentence.replace('choicebe','choice be')
    sentence = sentence.replace('choicebring','choice bring')
    sentence = sentence.replace('choicebut','choice but')
    sentence = sentence.replace('choicesand','choices and')
    sentence = sentence.replace('choicesbasically','choices basically')
    sentence = sentence.replace('choicesok','choices ok')
    sentence = sentence.replace('choicespersonally','choices personally')
    sentence = sentence.replace('choicesvalue','choices value')
    sentence = sentence.replace('choises','choices')
    sentence = sentence.replace('choosejust','choose just')
    sentence = sentence.replace('citycrowded','city crowded')
    sentence = sentence.replace('cleanlinessit','cleanliness it')
    sentence = sentence.replace('cleanmodern','clean modern')
    sentence = sentence.replace('cleanneed','clean need')
    sentence = sentence.replace('cleanwe','clean we')
    sentence = sentence.replace('closedyes','closed yes')
    sentence = sentence.replace('coldiced','cold iced')
    sentence = sentence.replace('conditionedthere','conditioned there')
    sentence = sentence.replace('conditionso','condition se')
    sentence = sentence.replace('courteateries','court eateries')
    sentence = sentence.replace('courthawker','court hawker')
    sentence = sentence.replace('courtshawker','courts hawker')
    sentence = sentence.replace('courtit','court it')
    sentence = sentence.replace('courtplenty','court plenty')
    sentence = sentence.replace('courtno','court no')
    sentence = sentence.replace('courtsingapore','court singapore')
    sentence = sentence.replace('crabmusselkangkong','crab mussel kang kong')
    sentence = sentence.replace('cravingssatisfied','cravings satisfied')
    sentence = sentence.replace('croissantsevery','croissants every')
    sentence = sentence.replace('cuisinegreat','cuisine great')
    sentence = sentence.replace('cuisinehow','cuisine how')
    sentence = sentence.replace('cuisinesfun','cuisines fun')
    sentence = sentence.replace('cuisinesgoing','cuisines going')
    sentence = sentence.replace('cuisinesthe','cuisines the')
    sentence = sentence.replace('cuisinesthey','cuisines they')
    sentence = sentence.replace('cuisineswe','cuisines we')
    sentence = sentence.replace('cuminspicy','cumin spicy')
    sentence = sentence.replace('customersdont','customers do not')
    sentence = sentence.replace('customershowever','customers however')
    sentence = sentence.replace('daytimeeverything','daytime everything')
    sentence = sentence.replace('definately','definitely')
    sentence = sentence.replace('deffo','definitely')
    sentence = sentence.replace('definetely','definitely')
    sentence = sentence.replace('delicaciesof','delicacies of')
    sentence = sentence.replace('deliciousafter','delicious after')
    sentence = sentence.replace('deliciouscant','delicious cannot')
    sentence = sentence.replace('deliciousi','delicious i')
    sentence = sentence.replace('deliciousin','delicious in')
    sentence = sentence.replace('deliciouswell','delicious well')
    sentence = sentence.replace('delicius','delicious')
    sentence = sentence.replace('deliscious','delicious')
    sentence = sentence.replace('desertedi','deserted i')
    sentence = sentence.replace('desertssatay','deserts satay')
    sentence = sentence.replace('dinnerchoice','dinner choice')
    sentence = sentence.replace('dinnerjugs','dinner jugs')
    sentence = sentence.replace('dinnersince','dinner since')
    sentence = sentence.replace('dinnerthe','dinner the')
    sentence = sentence.replace('dinnerwe','dinner we')
    sentence = sentence.replace('diorreah','diarrhoea')
    sentence = sentence.replace('disappointmentthis','disappointment this')
    sentence = sentence.replace('dishesdesserts','dishes desserts')
    sentence = sentence.replace('dishesprices','dishes prices')
    sentence = sentence.replace('dishhighly','dish highly')
    sentence = sentence.replace('dissapointed','disappointed')
    sentence = sentence.replace('dissapointedthe','disappointed the')
    sentence = sentence.replace('dissapointmentas','disappointedment as')
    sentence = sentence.replace('dissappointed','disappointed')
    sentence = sentence.replace('drinksunfortunately','drinks unfortunately')
    sentence = sentence.replace('drinktoo','drink too')
    sentence = sentence.replace('eatdrink','eat drink')
    sentence = sentence.replace('eatenwe','eaten we')
    sentence = sentence.replace('eatfor','eat for')
    sentence = sentence.replace('eatingas','eating as')
    sentence = sentence.replace('enormouswith','enormous with')
    sentence = sentence.replace('enoughsatay','enough satay')
    sentence = sentence.replace('eveningeat','evening eat')
    sentence = sentence.replace('eveninggreat','evening great')
    sentence = sentence.replace('eveningsnote','evenings note')
    sentence = sentence.replace('excellentand','excellent and ')
    sentence = sentence.replace('excellentbest','excellent best')
    sentence = sentence.replace('excellentwe','excellent we')
    sentence = sentence.replace('excllent','excellent')
    sentence = sentence.replace('execellent','excellent')
    sentence = sentence.replace('experance','experience')
    sentence = sentence.replace('experiencea','experience a')
    sentence = sentence.replace('experiencethe','experience the')
    sentence = sentence.replace('experiencereasonably','experience reasonably')
    sentence = sentence.replace('experienctip','experience tip')
    sentence = sentence.replace('experiencewe','experience we')
    sentence = sentence.replace('experince','experience')
    sentence = sentence.replace('fantasticgreat','fantastic great')
    sentence = sentence.replace('fantasticrecommend','fantastic recommend')
    sentence = sentence.replace('fasterbetter','faster better')
    sentence = sentence.replace('fastspeedy','fast speedy')
    sentence = sentence.replace('favorite','favourite')
    sentence = sentence.replace('favorite','favourite')
    sentence = sentence.replace('favouritesfair','favourites fair')
    sentence = sentence.replace('favouritesthis','favourites this')
    sentence = sentence.replace('favouriteswe','favourites we')
    sentence = sentence.replace('foodcuisines','food cuisines')
    sentence = sentence.replace('foodenjoyable','food enjoyable')
    sentence = sentence.replace('foodfantastic','food fantastics')
    sentence = sentence.replace('foodgreat','food great')
    sentence = sentence.replace('hereeverything','here everything')
    sentence = sentence.replace('herefantastic','here fantastic')
    sentence = sentence.replace('insidewaitingwaitingagain','inside waiting waiting again')
    sentence = sentence.replace('lovelly','lovely')
    sentence = sentence.replace('imo','in my opinion')
    sentence = sentence.replace('lunchdinner','lunch dinner')
    sentence = sentence.replace('lunchi','lunch i')
    sentence = sentence.replace('lunchso','lunch so')
    sentence = sentence.replace('lunchthe','lunch the')
    sentence = sentence.replace('lunchthere','lunch there')
    sentence = sentence.replace('lunchvariety','lunch variety')
    sentence = sentence.replace('mannersdefinitely','manners definitely')
    sentence = sentence.replace('marketconveniently','market conveniently')
    sentence = sentence.replace('mustavailable','must available')
    sentence = sentence.replace('mustdo','must do')
    sentence = sentence.replace('mustsee','must see')
    sentence = sentence.replace('mustseeexperience','must see experience')
    sentence = sentence.replace('musttry','must try')
    sentence = sentence.replace('mustvisit','must visit')
    sentence = sentence.replace('noodlesat','noodles at')
    sentence = sentence.replace('noodlesfrom','noodles from')
    sentence = sentence.replace('noodlesmango','noodles mango')
    sentence = sentence.replace('nottobemissed','not to be missed')
    sentence = sentence.replace('nutshellambience','nutshell ambience')
    sentence = sentence.replace('peopleas','people as')
    sentence = sentence.replace('peoplehaving','people having')
    sentence = sentence.replace('peoplepersonally','people personally')
    sentence = sentence.replace('peoplesuch','people such')
    sentence = sentence.replace('peoplethere','people there')
    sentence = sentence.replace('peoplewe','people we')
    sentence = sentence.replace('peoplewatch','people watch')
    sentence = sentence.replace('peoplewell','people well')
    sentence = sentence.replace('quicklyhopefully','quickly hopefully')
    sentence = sentence.replace('quicklynext','quickly next')
    sentence = sentence.replace('racommended','recommended')
    sentence = sentence.replace('ratedisasterfor','rate disaster for')
    sentence = sentence.replace('ricedelicious','rice delicious')
    sentence = sentence.replace('ricebest','rice best')
    sentence = sentence.replace('wellcheck','well check')
    sentence = sentence.replace('wellchinese','well chinese')
    sentence = sentence.replace('wellhighly','well highly')
    sentence = sentence.replace('wellin','well in')
    sentence = sentence.replace('wellknown','well known')
    sentence = sentence.replace('wellmany','well many')
    sentence = sentence.replace('wellmarinated','well marinated')
    sentence = sentence.replace('welloiled','well oiled')
    sentence = sentence.replace('wellresearched','well researched')
    sentence = sentence.replace('wellserviced','well serviced')
    sentence = sentence.replace('wellstall','well stall')
    sentence = sentence.replace('wellthe','well the')
    sentence = sentence.replace('wellthis','well this')
    sentence = sentence.replace('wellventilated','well ventilated')
    sentence = sentence.replace('yummm','yummy')    
    sentence = sentence.replace('satays','satay')    
    sentence = sentence.replace('tables','table')
    sentence = sentence.replace('tastes','taste')
    
    return sentence
clean = lambda x : clean_rd_2(x)
df.review = df['review'].apply(clean)
df2.review = df2['review'].apply(clean)

df

import pickle
# after round 2 of cleaning
df.to_pickle("new_df.pkl")
df2.to_pickle("new_df2.pkl")
